Error message for SSN updated and different cases included-Broker role only
When unknown user is selected but known SSN is entered a popup message
should show up asking the user if  he/she wants to use DB data. If "yes"
the unknown user checkbox should be switched to known and data should be
fetched from DB (not shown on the front-end). If "no" is selected the DB
table will be updated after the user enters all the mandatory fields.