Error handling on TransactionAccount on FNOL
While entering FNOL - add error handling so that when a non-regular bank
account is entered (for example 2130202051128468) message pops up - "The
bank account you have entered is not a valid Macedonian bank account. Do
you want to correct it or switch to Foreign bank account" followed by
the buttons "Correct entry" and "Select Foreign bank account". If Select
bank account is clicked the Foreign bank account checkbox should be
checked and the courser should be focused on the Bank name.